fix: use of docker testcontainers for neo4j bolt testing
