+++ neo4k-api/src/test/kotlin/me/roybailey/neo4k/server/TestApiServer.kt
@@ -1,24 +1,40 @@
 package me.roybailey.neo4k.server
 
 import io.javalin.Javalin
+import org.testcontainers.Testcontainers
 import java.net.ServerSocket
+import java.util.concurrent.ConcurrentHashMap
 
 
-class TestApiServer {
+class TestApiServer(val name: String, val port: Int) {
 
     companion object {
+
         fun findFreePort(): Int {
             val serverSocket = ServerSocket(0)
             val port = serverSocket.localPort
             serverSocket.close()
             return port
         }
+
+        private val hashServers = ConcurrentHashMap<String, TestApiServer>()
+
+        fun createTestApiServer(name: String = "default", port: Int = findFreePort()): TestApiServer {
+            if (!hashServers.containsKey(name))
+                hashServers[name] = TestApiServer(name, port)
+            return testApiServer(name)
+        }
+
+        fun testApiServer(name: String = "default"): TestApiServer {
+            return hashServers[name]!!
+        }
+
+        fun getTestApiServers(): List<TestApiServer> = hashServers.values.toList()
     }
 
-    lateinit var app:Javalin
-    val port = findFreePort()
+    lateinit var app: Javalin
 
-    fun start(data:Any) {
+    fun start(data: Any) {
         app = Javalin.create().start(port)
         app.get("/testdata") { ctx ->
             ctx.json(data)
@@ -30,4 +46,5 @@ class TestApiServer {
     }
 
     val url: String = "http://localhost:$port"
+
 }
