+++ neo4k-api/src/test/resources/neo4j/import/products-100.csv
@@ -1,0 +1,101 @@
+ProductId,SupplierId,Name,Category,Units,Clearance,BestBefore
+PRD0000000001,SUP0000000001,Product 1,TOYS,100,true,2019-08-13
+PRD0000000002,SUP0000000001,Product 2,BOOKS,1000,false,2019-08-23
+PRD0000000003,SUP0000000001,Product 3,DIY,10,true,2019-08-03
+PRD0000000004,SUP0000000001,Product 4,FOOD,100,false,2019-08-13
+PRD0000000005,SUP0000000001,Product 5,TOYS,1000,true,2019-08-23
+PRD0000000006,SUP0000000001,Product 6,BOOKS,10,false,2019-08-03
+PRD0000000007,SUP0000000001,Product 7,DIY,100,true,2019-08-13
+PRD0000000008,SUP0000000001,Product 8,FOOD,1000,false,2019-08-23
+PRD0000000009,SUP0000000001,Product 9,TOYS,10,true,2019-08-03
+PRD0000000010,SUP0000000001,Product 10,BOOKS,100,false,2019-08-13
+PRD0000000011,SUP0000000002,Product 11,DIY,1000,true,2019-08-23
+PRD0000000012,SUP0000000002,Product 12,FOOD,10,false,2019-08-03
+PRD0000000013,SUP0000000002,Product 13,TOYS,100,true,2019-08-13
+PRD0000000014,SUP0000000002,Product 14,BOOKS,1000,false,2019-08-23
+PRD0000000015,SUP0000000002,Product 15,DIY,10,true,2019-08-03
+PRD0000000016,SUP0000000002,Product 16,FOOD,100,false,2019-08-13
+PRD0000000017,SUP0000000002,Product 17,TOYS,1000,true,2019-08-23
+PRD0000000018,SUP0000000002,Product 18,BOOKS,10,false,2019-08-03
+PRD0000000019,SUP0000000002,Product 19,DIY,100,true,2019-08-13
+PRD0000000020,SUP0000000002,Product 20,FOOD,1000,false,2019-08-23
+PRD0000000021,SUP0000000003,Product 21,TOYS,10,true,2019-08-03
+PRD0000000022,SUP0000000003,Product 22,BOOKS,100,false,2019-08-13
+PRD0000000023,SUP0000000003,Product 23,DIY,1000,true,2019-08-23
+PRD0000000024,SUP0000000003,Product 24,FOOD,10,false,2019-08-03
+PRD0000000025,SUP0000000003,Product 25,TOYS,100,true,2019-08-13
+PRD0000000026,SUP0000000003,Product 26,BOOKS,1000,false,2019-08-23
+PRD0000000027,SUP0000000003,Product 27,DIY,10,true,2019-08-03
+PRD0000000028,SUP0000000003,Product 28,FOOD,100,false,2019-08-13
+PRD0000000029,SUP0000000003,Product 29,TOYS,1000,true,2019-08-23
+PRD0000000030,SUP0000000003,Product 30,BOOKS,10,false,2019-08-03
+PRD0000000031,SUP0000000004,Product 31,DIY,100,true,2019-08-13
+PRD0000000032,SUP0000000004,Product 32,FOOD,1000,false,2019-08-23
+PRD0000000033,SUP0000000004,Product 33,TOYS,10,true,2019-08-03
+PRD0000000034,SUP0000000004,Product 34,BOOKS,100,false,2019-08-13
+PRD0000000035,SUP0000000004,Product 35,DIY,1000,true,2019-08-23
+PRD0000000036,SUP0000000004,Product 36,FOOD,10,false,2019-08-03
+PRD0000000037,SUP0000000004,Product 37,TOYS,100,true,2019-08-13
+PRD0000000038,SUP0000000004,Product 38,BOOKS,1000,false,2019-08-23
+PRD0000000039,SUP0000000004,Product 39,DIY,10,true,2019-08-03
+PRD0000000040,SUP0000000004,Product 40,FOOD,100,false,2019-08-13
+PRD0000000041,SUP0000000005,Product 41,TOYS,1000,true,2019-08-23
+PRD0000000042,SUP0000000005,Product 42,BOOKS,10,false,2019-08-03
+PRD0000000043,SUP0000000005,Product 43,DIY,100,true,2019-08-13
+PRD0000000044,SUP0000000005,Product 44,FOOD,1000,false,2019-08-23
+PRD0000000045,SUP0000000005,Product 45,TOYS,10,true,2019-08-03
+PRD0000000046,SUP0000000005,Product 46,BOOKS,100,false,2019-08-13
+PRD0000000047,SUP0000000005,Product 47,DIY,1000,true,2019-08-23
+PRD0000000048,SUP0000000005,Product 48,FOOD,10,false,2019-08-03
+PRD0000000049,SUP0000000005,Product 49,TOYS,100,true,2019-08-13
+PRD0000000050,SUP0000000005,Product 50,BOOKS,1000,false,2019-08-23
+PRD0000000051,SUP0000000006,Product 51,DIY,10,true,2019-08-03
+PRD0000000052,SUP0000000006,Product 52,FOOD,100,false,2019-08-13
+PRD0000000053,SUP0000000006,Product 53,TOYS,1000,true,2019-08-23
+PRD0000000054,SUP0000000006,Product 54,BOOKS,10,false,2019-08-03
+PRD0000000055,SUP0000000006,Product 55,DIY,100,true,2019-08-13
+PRD0000000056,SUP0000000006,Product 56,FOOD,1000,false,2019-08-23
+PRD0000000057,SUP0000000006,Product 57,TOYS,10,true,2019-08-03
+PRD0000000058,SUP0000000006,Product 58,BOOKS,100,false,2019-08-13
+PRD0000000059,SUP0000000006,Product 59,DIY,1000,true,2019-08-23
+PRD0000000060,SUP0000000006,Product 60,FOOD,10,false,2019-08-03
+PRD0000000061,SUP0000000007,Product 61,TOYS,100,true,2019-08-13
+PRD0000000062,SUP0000000007,Product 62,BOOKS,1000,false,2019-08-23
+PRD0000000063,SUP0000000007,Product 63,DIY,10,true,2019-08-03
+PRD0000000064,SUP0000000007,Product 64,FOOD,100,false,2019-08-13
+PRD0000000065,SUP0000000007,Product 65,TOYS,1000,true,2019-08-23
+PRD0000000066,SUP0000000007,Product 66,BOOKS,10,false,2019-08-03
+PRD0000000067,SUP0000000007,Product 67,DIY,100,true,2019-08-13
+PRD0000000068,SUP0000000007,Product 68,FOOD,1000,false,2019-08-23
+PRD0000000069,SUP0000000007,Product 69,TOYS,10,true,2019-08-03
+PRD0000000070,SUP0000000007,Product 70,BOOKS,100,false,2019-08-13
+PRD0000000071,SUP0000000008,Product 71,DIY,1000,true,2019-08-23
+PRD0000000072,SUP0000000008,Product 72,FOOD,10,false,2019-08-03
+PRD0000000073,SUP0000000008,Product 73,TOYS,100,true,2019-08-13
+PRD0000000074,SUP0000000008,Product 74,BOOKS,1000,false,2019-08-23
+PRD0000000075,SUP0000000008,Product 75,DIY,10,true,2019-08-03
+PRD0000000076,SUP0000000008,Product 76,FOOD,100,false,2019-08-13
+PRD0000000077,SUP0000000008,Product 77,TOYS,1000,true,2019-08-23
+PRD0000000078,SUP0000000008,Product 78,BOOKS,10,false,2019-08-03
+PRD0000000079,SUP0000000008,Product 79,DIY,100,true,2019-08-13
+PRD0000000080,SUP0000000008,Product 80,FOOD,1000,false,2019-08-23
+PRD0000000081,SUP0000000009,Product 81,TOYS,10,true,2019-08-03
+PRD0000000082,SUP0000000009,Product 82,BOOKS,100,false,2019-08-13
+PRD0000000083,SUP0000000009,Product 83,DIY,1000,true,2019-08-23
+PRD0000000084,SUP0000000009,Product 84,FOOD,10,false,2019-08-03
+PRD0000000085,SUP0000000009,Product 85,TOYS,100,true,2019-08-13
+PRD0000000086,SUP0000000009,Product 86,BOOKS,1000,false,2019-08-23
+PRD0000000087,SUP0000000009,Product 87,DIY,10,true,2019-08-03
+PRD0000000088,SUP0000000009,Product 88,FOOD,100,false,2019-08-13
+PRD0000000089,SUP0000000009,Product 89,TOYS,1000,true,2019-08-23
+PRD0000000090,SUP0000000009,Product 90,BOOKS,10,false,2019-08-03
+PRD0000000091,SUP0000000010,Product 91,DIY,100,true,2019-08-13
+PRD0000000092,SUP0000000010,Product 92,FOOD,1000,false,2019-08-23
+PRD0000000093,SUP0000000010,Product 93,TOYS,10,true,2019-08-03
+PRD0000000094,SUP0000000010,Product 94,BOOKS,100,false,2019-08-13
+PRD0000000095,SUP0000000010,Product 95,DIY,1000,true,2019-08-23
+PRD0000000096,SUP0000000010,Product 96,FOOD,10,false,2019-08-03
+PRD0000000097,SUP0000000010,Product 97,TOYS,100,true,2019-08-13
+PRD0000000098,SUP0000000010,Product 98,BOOKS,1000,false,2019-08-23
+PRD0000000099,SUP0000000010,Product 99,DIY,10,true,2019-08-03
+PRD0000000100,SUP0000000010,Product 100,FOOD,100,false,2019-08-13
