+++ pom.xml
@@ -29,6 +29,7 @@
         <junit.version>5.3.2</junit.version>
         <assertj.version>3.6.1</assertj.version>
         <mokito.version>2.1.0</mokito.version>
+        <testcontainers.version>1.11.4</testcontainers.version>
 
     </properties>
